requirements.txt
playwright
python-Wappalyzer
requests
beautifulsoup4